now also plotting error for the Malaga dataset

+++ src/featureMatching/KLT/getKLTMatches.m
@@ -16,11 +16,11 @@ initialize(KLT_tracker, keypoints_1, img0);
 % Track the points in the second image
 [keypoints_2, isMatched] = KLT_tracker(img1);
 
-matched_keypoints_1 = keypoints_1(isMatched, :);
+matched_keypoints_1 = double(keypoints_1(isMatched, :));
 if (withRounding)
-    matched_keypoints_2 = round(keypoints_2(isMatched, :)); % It seems like IP is not solved with rounding :'( !!!!
+    matched_keypoints_2 = double(round(keypoints_2(isMatched, :))); % It seems like IP is not solved with rounding :'( !!!!
 else
-    matched_keypoints_2 = keypoints_2(isMatched, :); 
+    matched_keypoints_2 = double(keypoints_2(isMatched, :)); 
 end
 end
